src/devscape/requirements.txt
astroid==3.3.11
black==25.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dill==0.4.0
idna==3.10
iniconfig==2.1.0
isort==6.1.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
pygame==2.6.1
Pygments==2.19.2
pylint==3.3.9
pytest==8.4.2
pytokens==0.1.10
requests==2.32.5
tomlkit==0.13.3
urllib3==2.5.0
mypy==1.8.0
types-requests==2.32.4.20250913
bandit==1.7.8
mutmut==2.4.0
pyyaml
jsonschema